fix: use u128 instead of i64 for latency measurements

+++ crates/router/src/routes/app.rs
@@ -56,8 +56,8 @@ pub trait AppStateInfo {
     fn add_merchant_id(&mut self, merchant_id: Option<String>);
     fn add_flow_name(&mut self, flow_name: String);
     fn get_request_id(&self) -> Option<String>;
-    fn add_external_call_latencies(&mut self, url_path: String, latency: i64);
-    fn get_external_call_latencies(&self) -> Option<i64>;
+    fn add_external_call_latencies(&mut self, url_path: String, latency: u128);
+    fn get_external_call_latencies(&self) -> Option<u128>;
 }
 
 impl AppStateInfo for AppState {
@@ -83,11 +83,11 @@ impl AppStateInfo for AppState {
     fn get_request_id(&self) -> Option<String> {
         self.api_client.get_request_id()
     }
-    fn add_external_call_latencies(&mut self, url_path: String, latency: i64) {
+    fn add_external_call_latencies(&mut self, url_path: String, latency: u128) {
         self.api_client
             .add_external_call_latencies(url_path, latency);
     }
-    fn get_external_call_latencies(&self) -> Option<i64> {
+    fn get_external_call_latencies(&self) -> Option<u128> {
         self.api_client.get_external_call_latencies()
     }
 }
